# 2024_05_20/mvc_dulceria/app/static/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Dulceria API",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "Dulces",
      "description": "Operaciones relacionadas con los dulces en la dulceria"
    }
  ],
  "paths": {
    "/api/candy": {
      "get": {
        "summary": "Obtiene la lista de todos los dulces",
        "security": [
          {
            "JWTAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de dulces",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Candy"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crea un nuevo dulce",
        "security": [
          {
            "JWTAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Candy"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Dulce creado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Candy"
                }
              }
            }
          }
        }
      }
    },
    "/api/candy/{id}": {
      "get": {
        "summary": "Obtiene un dulce por su ID",
        "security": [
          {
            "JWTAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles del dulce",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dulce"
                }
              }
            }
          },
          "404": {
            "description": "Dulce no encontrado"
          }
        }
      },
      "put": {
        "summary": "Actualiza un dulce por su id",
        "security": [
          {
            "JWTAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Candy"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Dulce actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Candy"
                }
              }
            }
          },
          "404": {
            "description": "Dulce no encontrado"
          }
        }
      },
      "delete": {
        "summary": "Elimina un dulce por su id",
        "security": [
          {
            "JWTAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Dulce eliminado"
          },
          "404": {
            "description": "Dulce no encontrado"
          }
        }
      }
    },

    "/api/users": {
      "get": {
        "summary": "Obtiene la lista de todos los usuarios",
        "security": [
          {
            "JWTAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuarios",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Registra un nuevo usuario",
        "security": [
          {
            "JWTAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado"
          },
          "400": {
            "description": "Solicitud incorrecta"
          }
        }
      }
    },
    "/api/users/login": {
      "post": {
        "summary": "Inicia sesión con un usuario existente",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inicio de sesión exitoso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string",
                      "description": "Token de acceso JWT para el usuario"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Credenciales inválidas"
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "summary": "Obtiene un usuario por su ID",
        "security": [
          {
            "JWTAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles del usuario",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        }
      },
      "put": {
        "summary": "Actualiza un usuario por su id",
        "security": [
          {
            "JWTAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        }
      },
      "delete": {
        "summary": "Elimina un dulce por su id",
        "security": [
          {
            "JWTAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Usuario eliminado"
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        }
      }
    }
    
  },
  "components": {
    "securitySchemes": {
      "JWTAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Candy": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "readOnly": true
          },
          "marca": {
            "type": "string"
          },
          "peso": {
            "type": "float"
          },
          "sabor": {
            "type": "string"
          },
          "origen": {
            "type": "string"
          }
        },
        "required": ["marca", "peso", "sabor", "origen"]
      },
      "User": {
        "type": "object",
        "properties": {
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["admin", "user"]
            }
          }
        },
        "required": ["first_name","last_name","username", "password"]
      },
      "Login": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": ["username", "password"]
      }
    }
  }
}